refactor: rename the application name and all files containing the old name
