+++ pubspec.yaml
@@ -1,4 +1,4 @@
-name: myschool
+name: studle
 description: "Application for students to research their data and performance at school"
 publish_to: "none"
