+++ test/mocks/calendar_day_off_entity_mock.dart
@@ -1,5 +1,5 @@
-import 'package:myschool/src/core/extension_types/guid.dart';
-import 'package:myschool/src/modules/calendar/domain/entities/calendar_day_off_entity.dart';
+import 'package:studle/src/core/extension_types/guid.dart';
+import 'package:studle/src/modules/calendar/domain/entities/calendar_day_off_entity.dart';
 
 final entity = CalendarDayOffEntity(
   id: Guid.empty(),
