+++ test/mocks/login_with_student_registration_mock.dart
@@ -1,4 +1,4 @@
-import 'package:myschool/src/modules/authentication/domain/dtos/login_with_student_registration_dto.dart';
+import 'package:studle/src/modules/authentication/domain/dtos/login_with_student_registration_dto.dart';
 
 const dto = LoginWithStudentRegistrationDto(
   registration: '00123456',
