+++ test/src/core/helpers/date_helper_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:intl/date_symbol_data_local.dart';
 import 'package:intl/intl.dart';
-import 'package:myschool/src/core/helpers/date_helper.dart';
+import 'package:studle/src/core/helpers/date_helper.dart';
 
 void main() {
   setUpAll(() async {
