+++ test/src/modules/authentication/domain/dtos/login_with_student_registration_dto_test.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 
 import 'package:flutter_test/flutter_test.dart';
-import 'package:myschool/src/modules/authentication/domain/dtos/login_with_student_registration_dto.dart';
+import 'package:studle/src/modules/authentication/domain/dtos/login_with_student_registration_dto.dart';
 
 import '../../../../../mocks/login_with_student_registration_mock.dart' as mock;
