+++ test/src/modules/calendar/domain/entities/calendar_day_off_entity_test.dart
@@ -1,8 +1,8 @@
 import 'dart:convert';
 
 import 'package:flutter_test/flutter_test.dart';
-import 'package:myschool/src/core/extension_types/guid.dart';
-import 'package:myschool/src/modules/calendar/domain/entities/calendar_day_off_entity.dart';
+import 'package:studle/src/core/extension_types/guid.dart';
+import 'package:studle/src/modules/calendar/domain/entities/calendar_day_off_entity.dart';
 
 import '../../../../../mocks/calendar_day_off_entity_mock.dart' as mock;
